LayoutTests/fast/js/postfix-syntax-expected.txt

This test checks whether various forms of postincrement expression are allowed. It should print the numbers 0-9 on separate lines, and nothing else (besides this paragraph).


0
1
2
3
4
5
6
7
8
9
